figuring out why I am having connection issues, but everything else I fixed

+++ back-end/src/db/migrations/20240603202532_create_tables.js
@@ -12,6 +12,9 @@ exports.up = function (knex) {
   });
 };
 
-exports.down = function (knex) {
-  return knex.schema.dropTable("tables");
+exports.down = async function (knex) {
+  const hasTablesTable = await knex.schema.hasTable("tables");
+  if (hasTablesTable) {
+    return knex.schema.dropTable("tables");
+  }
 };
